requirements.txt
Flask==2.2.2
tensorflow==2.10.0
Pillow==9.0.1
numpy==1.21.2